server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
Pillow
python-dotenv
